requirements.txt
appdirs==1.4.0
cycler==0.10.0
enum34==1.1.6
google-api-python-client==1.6.2
google-auth==0.7.0
google-auth-httplib2==0.0.2
google-cloud-core==0.22.1
google-cloud-vision==0.22.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
matplotlib==2.0.0
numpy==1.12.0
oauth2client==4.0.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
protobuf==3.2.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
Flask==0.12
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-WTF==0.13.1
Flask-Mail
youtube-dl==2017.2.17
algoliasearch==1.11.2